POLIO-1794: refactor PaginatedModelViewset
- rename as EtlModelViewset
- Add default EtlPaginator, with a default `page_size` of 20 and a `max_page_size` of 1000
- Allow overriding paginator if it is a subclass of EtlPaginator, to make sure there's always a `page_size` and a `max_page_size`
- update tests
